+++ tests/TestCase/Model/Entity/BannersPositionTest.php
@@ -31,12 +31,41 @@ use MeCms\Model\Entity\BannersPosition;
 class BannersPositionTest extends TestCase
 {
     /**
+     * @var \MeCms\Model\Entity\BannersPosition
+     */
+    protected $BannersPosition;
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->BannersPosition = new BannersPosition;
+    }
+
+    /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        unset($this->BannersPosition);
+    }
+
+    /**
      * Test for `__construct()` method
      * @test
      */
     public function testConstruct()
     {
-        $this->assertInstanceOf('MeCms\Model\Entity\BannersPosition', new BannersPosition);
+        $this->assertInstanceOf('MeCms\Model\Entity\BannersPosition', $this->BannersPosition);
     }
 
     /**
@@ -46,10 +75,8 @@ class BannersPositionTest extends TestCase
      */
     public function testNoAccessibleProperties()
     {
-        $entity = new BannersPosition();
-
-        $this->assertFalse($entity->isAccessible('id'));
-        $this->assertFalse($entity->isAccessible('banner_count'));
-        $this->assertFalse($entity->isAccessible('modified'));
+        $this->assertFalse($this->BannersPosition->isAccessible('id'));
+        $this->assertFalse($this->BannersPosition->isAccessible('banner_count'));
+        $this->assertFalse($this->BannersPosition->isAccessible('modified'));
     }
 }